==> Model Training/Spam Text Module/0. Data_Preparation.ipynb <==
# SMS Spam Detection Using Machine Learning


---


## Practice Module: Pattern Recognition Systems (PRS)

## Group: 18

## Members:

Lim Jun Ming, A0231523U

Mediana, A0231458E

Yeong Wee Ping, A0231533R

# Data Preparation

## 0. File Path & Library Setup

# Load All Necessary Packages

import os
from google.colab import drive

import pandas as pd
import numpy as np
import sklearn as sk
from sklearn.utils import shuffle

seed = 18

print("Versions of key libraries")
print("-------------------------")
print("pandas:  ", pd.__version__)
print("numpy:   ", np.__version__)
print("sklearn: ", sk.__version__)


# Mounting to Google Drive
drive.mount('/content/gdrive')

# Change Working Directory
os.chdir('/content/gdrive/My Drive/iss/prs_pm/training')

print('Working Directory: ')
!pwd

## 1. Load and Combine Raw Data

# Load raw data files
header = ["Label", "Text"]

raw1 = pd.read_csv("raw_data/SMSSpamCollection.csv", encoding="UTF-8", names=header)
raw2 = pd.read_csv("raw_data/SpamSMSCollected.csv", encoding="UTF-8", names=header)
raw3 = pd.read_csv("raw_data/KaggleSMSDataset.csv", encoding="UTF-8", names=header)

row1, clm1 = raw1.shape
row2, clm2 = raw2.shape
row3, clm3 = raw3.shape

print('Data file 1 has {:2d} rows of labelled data points'.format(row1))
print('-----------------------------------------------------------')
print(raw1.head())
print('-----------------------------------------------------------')
print(' ')
print('Data file 2 has {:2d} rows of labelled data points'.format(row2))
print('-----------------------------------------------------------')
print(raw2.head())
print('-----------------------------------------------------------')
print(' ')
print('Data file 3 has {:2d} rows of labelled data points'.format(row3))
print('-----------------------------------------------------------')
print(raw3.head())
print('-----------------------------------------------------------')

# Combine all 3 data files
raw = shuffle(pd.concat((raw1,raw2,raw3)), random_state=seed).reset_index(drop=True)

row, clm = raw.shape

print('Combined data file has {:2d} rows of labelled data points'.format(row))
print('-----------------------------------------------------------')
print(raw.head())
print('-----------------------------------------------------------')
print(' ')
print('Total of {:2d} data points labelled as ham'.format((raw['Label']=='ham').sum()))
print('Total of {:2d} data points labelled as spam'.format((raw['Label']=='spam').sum()))
print('Ratio of ham over spam labelled data points is {:.2f}'.format((raw['Label']=='ham').sum()/(raw['Label']=='spam').sum()))

## 2. Save and Export Combined Data File to folder

# Save in CSV format (without headers)

raw.to_csv('structured_data/smsdata.csv', index=False, header=False)
